+++ CHANGELOG.md
@@ -8,6 +8,8 @@
   - Copy Name - Copy full path to clipboard
   - Show in Viewer - Shows signal in viewer
 - Added tooltips to netlist view to show details on netlist elements
+- Setting a marker populates the time in the search bar
+- Added Auto Touchpad Scrolling mode
 
 ## 1.3.1 - Latest Release